+++ CHANGELOG.md
@@ -14,7 +14,15 @@ and this project adheres to
   rest api.
 - Commit analysis to be done on every `github_pull_request` regardless of if the
   `analyzeCommitApproval` config variable is set or not.
-- Removed `analyzeCommitApproval` config variable.
+
+### Removed
+
+- `analyzeCommitApproval` config variable.
+
+### Added
+
+- Retrying "Secondary Rate Limit" errors on graphQL queries.
+- Better logging of graphQL queries.
 
 ## 1.3.0 - 2021-08-27